fix: pin load_png_rgb axis order and shape-guard fallback with tests
Adds regression coverage the review flagged as missing: a real
pygame Surface->PNG->load_png_rgb round trip with W != H catches an
axis-swap regression by shape and asymmetric pixel values, and two
new background() tests drive a greyscale (ndim==2) and an RGBA
(H,W,4) override through the real _require_rgb guard, asserting the
fallback, the logged-once warning, and the failures entry.

Also matches background()'s override cast to palette()'s existing
.astype(np.uint8) defensive cast (copy=False to preserve the existing
by-identity contract for already-uint8 arrays), and documents in a
code comment why an absent override intentionally falls back without
logging, unlike an unreadable or invalid one.

+++ PyAitD/asset_resolver.py
@@ -49,6 +49,10 @@ class AssetResolver:
         if path in self._cache:
             return self._cache[path]
         if not path.is_file():
+            # A normal override directory leaves most floor/camera combinations
+            # un-overridden, so a plain absence is expected steady-state, not a
+            # failure: fall back silently, with no log and no failures entry.
+            # Only an override that exists but is unreadable or invalid logs.
             return None
         try:
             pixels = self._load_png(path)
@@ -67,7 +71,7 @@ class AssetResolver:
                 _require_rgb,
             )
             if pixels is not None:
-                return ImageAsset(pixels, True)
+                return ImageAsset(pixels.astype(np.uint8, copy=False), True)
         return ImageAsset(floor.camera_image(cam_idx), False)
 
     def palette(self, floor):
